requirements.txt
numpy
torch
torchvision
torchaudio
pillow
elasticsearch
pymilvus
faiss-cpu
